Starting Metadata from Scratch
Existing metadata wasn't actually able to deploy to a QA scratch org. After some testing and work, this version of the entire project metadata seems to actually be willing to deploy. Still requires that you activate the Lightning Record Pages for Case and for Contact in order to look like the app/org we want. But a huge improvement!