More parts of the app
Adds: home page, brand colors in OCC app, OCC logo, two reports for the home page. Adds Mopsy as a dependency and puts it on the Home page. There are post-install steps required: Activate Brand and Theme for OCC and set the Mopsy component on the homepage to see My Tasks.