Prerelease Refinements
Updates to repo to bring us back in line with what OCC team has developed in the original production org.
NOTE: My source_retrieve brought down the emailservices (email-to-case activation) as well as a handful of profiles that I know we don't want but I wasn't sure how to safely cut out. Also, I have added Path to the Case flexipage. For that to work there also had to be an activation of Path in the org and the creation of a pathAssistant. Not sure what of those will successfully deploy, so we should monitor. (Possibly the flexipage with Path component will deploy but the admin will have to create the Path as a manual post-install that will then just work upon reload?)